Move resolvers to services, test valid params
