+++ src/test/llm/llmSpecificTestUtils/transformParams.ts
@@ -1,31 +1,5 @@
-import { MultiroundProfile } from "../../../llm/llmServices/modelParams";
-import { UserMultiroundProfile } from "../../../llm/userModelParams";
-
 import { MockLLMModelParams } from "./mockLLMService";
-
-export interface UserModelParamsAddOns {
-    modelId?: string;
-    choices?: number;
-    systemPrompt?: string;
-    maxTokensToGenerate?: number;
-    tokensLimit?: number;
-    multiroundProfile?: UserMultiroundProfile;
-}
-
-export interface ModelParamsAddOns {
-    modelId?: string;
-    choices?: number;
-    systemPrompt?: string;
-    maxTokensToGenerate?: number;
-    tokensLimit?: number;
-    multiroundProfile?: MultiroundProfile;
-}
-
-export interface MultiroundProfileAddOns {
-    maxRoundsNumber?: number;
-    defaultProofFixChoices?: number;
-    proofFixPrompt?: string;
-}
+import { MultiroundProfileAddOns } from "./modelParamsAddOns";
 
 export function enhanceMockParams(
     basicMockParams: MockLLMModelParams,
